lcj bot/requirements.txt.txt
nonebot2
nonebot-adapter-onebot
requests